+++ src/start.js
@@ -1,11 +1,5 @@
 const { app, BrowserWindow, ipcMain } = require('electron')
-const path = require('path')
 const url = require('url')
-const { fork } = require('child_process');
-
-const {
-    events,
-} = require('./constants/index');
 
 let mainWindow
 
@@ -49,44 +43,8 @@ app.on('activate', () => {
     }
 })
 
-require('./ipc/files');
-
 /**
- * TODO: Put this in its own file in /ipc/
- * Kicks off the tobii.listen() loop.
- * 
- * Each time the eyetracking recognizes a new focus event,
- * the metadata of said event to emitted to the renderer process
- * using a event.reply() call. 
+ * Register IPC hooks.
  */
-let eyetrackingProcess;
-ipcMain.on(events.ASYNC_LISTEN, (event, arg) => {
-
-    // Check if there is a currently running eyetracking process
-    // if so, murder it
-    if (eyetrackingProcess) {
-        console.log(`Killing eyetracking process (${eyetrackingProcess.pid || 'no pid found'})`);
-        eyetrackingProcess.kill('SIGINT');
-    }
-
-    // fork a child process to run the eyetracking module
-    eyetrackingProcess = fork(path.join(__dirname, 'eyetracking.js'), [], {
-        stdio: ['pipe', 'pipe', 'pipe', 'ipc']
-    });
-
-    console.log(`Forking eyetracking process (${eyetrackingProcess.pid || 'no pid found'})`);
-    console.log(`(${eyetrackingProcess.pid}): w${arg.width} h${arg.height} r${arg.rectangles.length}`);
-
-    // Send the screen metadata to start the listen loop.
-    eyetrackingProcess.send(arg);
-
-    // When the forked process emits a message, push to the render process
-    eyetrackingProcess.on('message', (evt) => {
-        let payload = {
-            ...evt,
-            ...arg.rectangles[evt.id]
-        };
-
-        event.reply(events.ASYNC_GAZE_FOCUS_EVENT, payload);
-    });
-});
+require('./ipc/files');
+require('./ipc/eyetracking');
